refactor: extract function `loadParsedRules`
